Had not realized that amrex::Vector size() returned a Long

+++ Source/Soot.cpp
@@ -28,8 +28,9 @@ PeleC::addSootDerivePlotVars(
   // Add in soot variables
   amrex::Vector<std::string> sootNames = {"rho_soot", "sum_rho_soot"};
   derive_lst.add(
-    "soot_vars", amrex::IndexType::TheCellType(), sootNames.size(), sootNames,
-    soot_genvars, amrex::DeriveRec::TheSameBox);
+    "soot_vars", amrex::IndexType::TheCellType(),
+    static_cast<int>(sootNames.size()), sootNames, soot_genvars,
+    amrex::DeriveRec::TheSameBox);
   derive_lst.addComponent("soot_vars", desc_lst, State_Type, URHO, 1);
   derive_lst.addComponent(
     "soot_vars", desc_lst, State_Type, UFSOOT, NUM_SOOT_MOMENTS + 1);
@@ -38,8 +39,8 @@ PeleC::addSootDerivePlotVars(
   amrex::Vector<std::string> large_part_names = {"NL", "soot_V_L", "soot_S_L"};
   derive_lst.add(
     "soot_large_particles", amrex::IndexType::TheCellType(),
-    large_part_names.size(), large_part_names, soot_largeparticledata,
-    amrex::DeriveRec::TheSameBox);
+    static_cast<int>(large_part_names.size()), large_part_names,
+    soot_largeparticledata, amrex::DeriveRec::TheSameBox);
   derive_lst.addComponent(
     "soot_large_particles", desc_lst, State_Type, UFSOOT, NUM_SOOT_MOMENTS + 1);
 }
